rtc.h refactor
