+++ board/stm32fx/board.h
@@ -8,6 +8,7 @@
 #ifdef PANDA
   #include "drivers/fan.h"
   #include "drivers/rtc.h"
+  #include "stm32fx/llrtc.h"
   #include "drivers/clock_source.h"
   #include "boards/white.h"
   #include "boards/grey.h"
